pyproject.toml
[project]
name = "val-mal-bot"
version = "0.1.0"
description = "Valin Malach's bot for Twitch/Discord"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "asyncpg>=0.30.0",
    "atproto>=0.0.62",
    "attrs>=25.3.0",
    "audioop-lts>=0.2.2",
    "blinker>=1.9.0",
    "boto3>=1.40.32",
    "botocore>=1.40.32",
    "certifi>=2025.8.3",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.3",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "cryptography>=45.0.7",
    "discord>=2.3.2",
    "discord-py>=2.6.3",
    "dnspython>=2.8.0",
    "flask>=3.1.2",
    "frozenlist>=1.7.0",
    "h11>=0.16.0",
    "h2>=4.3.0",
    "hpack>=4.1.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "hypercorn>=0.17.3",
    "hyperframe>=6.1.0",
    "idna>=3.10",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "jmespath>=1.0.1",
    "libipld>=3.1.1",
    "mako>=1.3.10",
    "markupsafe>=3.0.2",
    "multidict>=6.6.4",
    "pandas>=2.3.2",
    "priority>=2.0.0",
    "propcache>=0.3.2",
    "pyarrow>=21.0.0",
    "pycparser>=2.23",
    "pydantic>=2.11.9",
    "pydantic-core>=2.33.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "quart>=0.20.0",
    "requests>=2.32.5",
    "rfc3986>=2.0.0",
    "s3transfer>=0.14.0",
    "sentry-sdk>=2.38.0",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "truststore>=0.10.4",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.1",
    "tzdata>=2025.2",
    "urllib3>=2.5.0",
    "websockets>=13.1",
    "werkzeug>=3.1.3",
    "wsproto>=1.2.0",
    "yarl>=1.20.1",
]
